pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ro-design-mcp"
version = "0.1.0"
description = "MCP server for reverse osmosis system design optimization"
readme = "README.md"
authors = [
    {name = "RO Design MCP Contributors"}
]
license = {text = "MIT"}
requires-python = ">=3.10"
dependencies = [
    "fastmcp>=0.1.0",
    "numpy>=1.24.0",
    "papermill>=2.5.0",
    "nbformat>=5.0.0",
    "pydantic>=2.0.0",
    "watertap>=0.11.0",
    "pyomo>=6.7.0",
    "idaes-pse>=2.2.0",
    "pandas>=2.0.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0"
]

[project.scripts]
ro-design-server = "server:main"

[tool.hatch.build.targets.wheel]
packages = ["server.py", "utils"]

[tool.ruff]
line-length = 88
target-version = "py310"
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]  # Line too long

[tool.black]
line-length = 88
target-version = ['py310']

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true